% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_knockoff.R
\name{adaptive_knockoff}
\alias{adaptive_knockoff}
\title{Adaptive knockoff}
\usage{
adaptive_knockoff(W, U, fdr = 0.1, offset = 1,
  method = filter_randomForest, reveal_prop = 0.5, mute = TRUE)
}
\arguments{
\item{U}{p-by-r vector of the side information.}

\item{fdr}{The target FDR level(s). The default is 0.1.}

\item{offset}{0 or 1 (default: 1).}

\item{method}{The method used for evaluating the probability of being null. The default is random forest.}

\item{w}{n-by-p martrix of the covariates.}
}
\description{
This function takes W as an input and outpus the selection set generated by the adaptive knockoff procedure
}
