% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_glmnet.R
\name{calculate.fdphat}
\alias{calculate.fdphat}
\title{Calculating fdphat with threshold tau}
\usage{
calculate.fdphat(W, tau, offset = 1)
}
\arguments{
\item{W}{importance statsitic}

\item{tau}{threshold}

\item{offset}{offset of fdphat, takes balue in {0,1}}
}
\description{
Calculating fdphat with threshold tau
}
\keyword{internal}
