% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_randomForest.R
\name{filter_randomForest}
\alias{filter_randomForest}
\title{Adaptive Knockoff Filter With Random Forest}
\usage{
filter_randomForest(W, z, alpha = 0.1, offset = 1, reveal_prop = 0.5,
  mute = TRUE)
}
\arguments{
\item{W}{vector of length p, denoting the imporatence statistics calculated by \code{\link[knockoff]{knockoff.filter}}.}

\item{z}{p-by-r matrix of side information.}

\item{alpha}{target FDR level (default is 0.1).}

\item{offset}{either 0 or 1 (default: 1). The offset used to compute the rejection threshold on the statistics. For details, see \code{\link[knockoff]{knockoff.threshold}}.}

\item{reveal_prop}{The proportion of hypotheses revealed at intialization (default is 0.5).}

\item{mute}{whether \eqn{\hat{fdp}} of each iteration is printed (defalt is TRUE).}
}
\value{
A list of the following:
 \item{nrejs}{The number of rejections for each specified target fdr (alpha) level}.
 \item{rejs}{Rejsction set fot each specified target fdr (alpha) level}.
 \item{rej.path}{The order of the hypotheses (used for diagnostics)}.
 \item{unrevealed.id}{id of the hypotheses that are nor revealed in the end (used for diagnostics)}.
 \item{tau}{Threshold of each target FDR level (used for diagnostics)}.
 \item{acc}{The accuracy of classfication at each step (used for diagnostics)}.
}
\description{
filter_glm returns a set of rejections with FDR controlled at custom target
}
\examples{
#Generating data
p=100;n=100;k=40;
mu = rep(0,p); Sigma = diag(p)
X = matrix(rnorm(n*p),n)
nonzero = 1:k
beta = 5*(1:p\%in\%nonzero)*sign(rnorm(p))/ sqrt(n)
y = X\%*\%beta + rnorm(n,1)

#Generate knockoff copy
Xk = create.gaussian(X,mu,Sigma)

#Gnerate importance statistic using knockoff package
W = stat.glmnet_coefdiff(X,Xk,y)

#Using filer_gam to obtain the final rejeciton set
z = 1:p #Use the location of the hypotheses as the side information
result = filter_glm(W,z)

}
\seealso{
Other filter: \code{\link{filter_EM}},
  \code{\link{filter_gam}}, \code{\link{filter_glmnet}},
  \code{\link{filter_glm}},
  \code{\link{filter_randomForest_getorder}}
}
\concept{filter}
