% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ordering.R
\name{plot_ordering}
\alias{plot_ordering}
\title{Diagnostics for adative knockoff results}
\usage{
plot_ordering(mdl, nonzero = NULL, start_index = 1, alpha = NULL)
}
\arguments{
\item{mdl}{the result given by adaptive knockoff filters.}

\item{nonzero}{The true signals (default is NULL).}
}
\value{
A plot of the realized ordering of hypotheses.
}
\description{
Plot_ordering takes the adaptive knockoff result as input and plot the ordering.
}
\examples{
#Generating data
p=100;n=100;k=40;
mu = rep(0,p); Sigma = diag(p)
X = matrix(rnorm(n*p),n)
nonzero = 1:k
beta = 5*(1:p\%in\%nonzero)*sign(rnorm(p))/ sqrt(n)
y = X\%*\%beta + rnorm(n,1)

#Generate knockoff copy
Xk = create.gaussian(X,mu,Sigma)

#Generate importance statistic using knockoff package
W = stat.glmnet_coefdiff(X,Xk,y)

#Using filer_EM to obtain the final rejeciton set
U = 1:p #Use the location of the hypotheses as the side information
result = filter_randomForest(W,U)
plot_ordering(result,nonzero = nonzero)
}
\seealso{
Other plot: \code{\link{plot_rank}},
  \code{\link{plot_vanilla}}
}
\concept{plot}
