% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vanilla.R
\name{plot_vanilla}
\alias{plot_vanilla}
\title{Diagnostics for adative knockoff results}
\usage{
plot_vanilla(W, alpha, nonzero = NULL, start_index = 1)
}
\arguments{
\item{W}{vector of length p, denoting the imporatence statistics calculated by \code{\link[knockoff]{knockoff.filter}}.}

\item{alpha}{A vector of all target FDR levels.}

\item{nonzero}{The true signals (default is NULL).}
}
\value{
A plot of the realized ordering by vanilla knockoff.
}
\description{
Plot_vanilla plots the ordering of vanilla knockoff as reference.
}
\seealso{
Other plot: 
\code{\link{plot_ordering}()},
\code{\link{plot_rank}()}
}
\concept{plot}
